+++ app/src/activity/server_list.cpp
@@ -89,7 +89,7 @@ public:
         return cell;
     }
 
-    void onItemSelected(brls::View* recycler, size_t index) override {
+    void onItemSelected(brls::Box* recycler, size_t index) override {
         brls::Application::blockInputs();
 
         brls::async([this, index]() {